requirements.txt
pytest==7.4.0
pytest-html==3.2.0
pytest-xdist==3.3.1
playwright==1.40.0
selenium==4.14.0
beautifulsoup4==4.12.2
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
browserstack-local==1.2.8
browserstack-sdk==1.0.0
allure-pytest==2.13.2
Faker==20.1.0
